+++ packages/fury-adapter-oas3-parser/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Fury OAS3 Parser Changelog
 
+## Master
+
+### Enhancements
+
+- Adds support for example values for text based media types such as
+  `text/*` and `application/xml`.
+
 ## 0.9.0 (2019-07-02)
 
 This release brings significant performance improvement to parsing large